Version 1.2
Kai's Commons is now a required library. With this change, SynX Sponge
version is now relying on the built-in HikariCP. Other versions are
using the extremely basic connection pool on Kai's Commons. This may be
replaced with a better version in a future version.
Getting the Object from a Packet on the onPacketReceived method is now
easier thanks to the Packet#getObject() and (T)
Packet#getObject(Class<T>) methods.